securely handle committee changes w/ process approvals:
Add lastCommitteeChange global variable
to keep track changes on the committee list.

When the committee is modified, creating a process
or updating it in the same block is not allowed.

If committee is modified, in the next tally approval,
the process approval list is updated accordingly
to the recent committee changes.
This is done, between others, by comparing the process security block
and the last block with committee changes.